Function documentation

NPDemand.define_problem — Function

define_problem(df::DataFrame; exchange = [], index_vars = ["prices"], FE = [], constraints = [], bO = 2, tol = 1e-5)

Constructs a problem::NPDProblem using the provided problem characteristics. Inputs:

exchange: A vector of groups of products which are exchangeable. E.g., with 4 goods, if the first

and second are exchangeable and so are the third and fourth, set exchange = [[1 2], [3 4]].

index_vars: String array listing column names in df which represent variables that enter the inverted index.

"prices" must be the first element of index_vars

FE: String array listing column names in df which should be included as fixed effects.

Note: All fixed effects are estimated as parameters by the minimizer, so be careful adding fixed effects for variables that take many values.

bO: Order of the univariate Bernstein polynomials in market shares. Default is 2.

constraint_tol: Tolerance specifying tightness of constraints

obj_xtol: Tolerance specifying x_tol in Optim.Options()

obj_ftol: Tolerance specifying f_tol in Optim.Options()

chunk_size: Controls chunk size in ForwardDiff.Gradient autodiff calculation for nonlinear constraints. Only used if :subs_in_group specified.

constraints: A list of symbols of accepted constraints. Currently supported constraints are:

:monotone

:all_substitutes

:diagonal_dominance_group

:diagonal_dominance_all

:subs_in_group (Note: this constraint is the only available nonlinear constraint and will slow down estimation considerably)

verbose: if false, will not print updates as problem is generated

NPDemand.update_constraints! — Function

update_constraints!(problem::NPDProblem, new_constraints::Vector{Symbol})

Re-calculates constraint matrices under new_constraints, ignoring previous constraints used to define the problem. The exchangeability constraint is an exception. To change anything about the structure of exchangeability for the problem changes, define a new problem.

NPDemand.estimate! — Function

estimate!(problem::NPDProblem; max_iterations = 10000, show_trace = false, chunk_size = Int[])

estimate! solves problem subject to provided constraints, and replaces problem.results with the resulting parameter vector. To enforce constraints, we iteratively minimize the sum of the objective function and a penalty term capturing violations of the constraint. For each outer iteration, we increase the size of the penalty term by an order of magnitude.

Options:

max_inner_iterations: controls the number of inner iterations for each call to the optimizer via Optim.jl

max_outer_iterations: controls the number of total calls to Optim, i.e., the number of times the penalty enforcing the constraints will increase before exiting

show_trace: if true, Optim will print the trace for each outer iteration.

verbose: if false, will not print outer iteration updates within optimization

NPDemand.price_elasticities! — Function

price_elasticities!(problem::NPDProblem)

Takes the solved problem as first argument, a DataFrame as the second argument, and evaluates all price elasticities in-sample. Currently does not calculate out-of-sample price elasticities. For this, use the function compute_demand_function!.

Results of this function are stored as a DataFrame in problem.allelasticities. Results can be summarized by hand or using the `summarizeelasticities` function. We also store the Jacobian of the demand function with respect to prices, which can be used to calculate other quantities of interest.

NPDemand.compute_demand_function! — Function

compute_demand_function!(problem, df; max_iter = 1000, show_trace = false)

compute_demand_function! estimates the demand function/curve using NPD estimates calculated via estimate!.

The function takes in an estimated problem::NPDProblem and a dataframe with counterfactual values of the covariates in the utility function. One must specify all fields that were used in estimation (including shares). The function will change the values of df[!,r"shares"] to take on the value of the estimated demand function.

Options:

max_iter: controls the number of iterations for the nonlinear solver calculating market shares. Default is 1000 but well-estimated problems should converge faster.

show_trace: if true, Optim will print the trace for each iteration of the nonlinear solver.

compute_own_elasticities: NOT yet implemented– if true, will also generate columns called own_elast which will include estimated own-price elasticities at all counterfactual points.